+++ projects/ngx-testing-tools/src/lib/common/test-bed/renderer/models/renderer-test-bed-options.model.ts
@@ -1,9 +1,12 @@
 import { SchemaMetadata } from '@angular/core';
-import { BaseTestBedOptions } from '../../base/models/custom-test-bed-options.model';
+import { BaseTestBedOptions } from '../../base/models/base-test-bed-options.model';
 
 export interface RendererTestBedOptions extends BaseTestBedOptions {
   /**
+   * The schema definitions.
    * @default []
+   * @see {@link CUSTOM_ELEMENTS_SCHEMA}
+   * @see {@link NO_ERRORS_SCHEMA}
    */
   schemas?: SchemaMetadata[];
   /**
