+++ projects/ngx-testing-tools/src/lib/module/test-bed/models/module-test-bed-options.model.ts
@@ -1,3 +1,3 @@
-import { BaseTestBedOptions } from '../../../common/test-bed/base/models/custom-test-bed-options.model';
+import { BaseTestBedOptions } from '../../../common/test-bed/base/models/base-test-bed-options.model';
 
 export interface ModuleTestBedOptions extends BaseTestBedOptions {}
