+++ projects/ngx-testing-tools/src/lib/service/test-bed/models/service-test-bed-options.model.ts
@@ -1,3 +1,3 @@
-import { BaseTestBedOptions } from '../../../common/test-bed/base/models/custom-test-bed-options.model';
+import { BaseTestBedOptions } from '../../../common/test-bed/base/models/base-test-bed-options.model';
 
 export interface ServiceTestBedOptions extends BaseTestBedOptions {}
